Make the terminal control also include input from SIGUSR1 and SIGUSR2, configurable via --ctrlusr1 and --ctrlusr2.
Actually, that is a hint to separate the control code from the terminal ... plain mpg123 can use those signals, too. Also this can be merged with the generic control and fifo sucking. But then ... are we coding a simple decoder ore a sophisticated media player? That's a question one might have posed ten years ago...