At least detect some out-of-disk situations in audio file writing.
But there are limits by our API choice to have no return value evaluation from wav_close and friends. Also I chose not to add deletion of created empty files. This would complicate the code, which currently does forget the file name very early, and also could result in undesired behaviour in case a file existed already (as a placeholder, for example) or if we are not dealing with a normal file at all.

I think mpg123 should not go around and try to delete files, even if it sometimes pointlessly created one.