libmpg123: reformat mpg123.h.in since the addition of MPG123_EXPORT prolonged things
Rationale: I'm giving up on my stance about editor programs of the future being
smarter dealing with breaking long lines and indent on display. We are in the
future and the programs still don't care (actually, it's partly the exact same
programs without change ...). I am starting to format things now with 80 chars
max, to have it human-readable.

But I keep using tabs for indent and (additional) spaces for alignment. So,
the 80-char-rule is depending on your tab width. I use 3. If you use wider
tabs, you should be prepared to use a wider window since you obviously
prefer things to be wide.